+++ NEWS.md
@@ -1,3 +1,12 @@
+## pipette 0.4.6 (2020-06-11)
+
+### Minor changes
+
+- `import`: Added support for `skip` argument, which allows the user to skip
+  a certain number of lines in the input.
+- `import` and `export` of source code lines now uses readr package internally
+  (`read_lines` and `write_lines`) instead of base `readLines` and `writeLines`.
+
 ## pipette 0.4.5 (2020-05-18)
 
 ### Minor changes